Clean up tendermint 35 files
