+++ packages/core/src/lib/tendermint35/encodings.ts
@@ -1,7 +1,7 @@
-import { toUtf8 } from "@cosmjs/encoding";
+import { toUtf8 } from '@cosmjs/encoding';
 
-import { ReadonlyDateWithNanoseconds } from "../dates";
-import { BlockId, Version } from "./responses";
+import { ReadonlyDateWithNanoseconds } from '@cosmjs/tendermint-rpc/build/dates';
+import { BlockId, Version } from './responses';
 
 /**
  * A runtime checker that ensures a given value is set (i.e. not undefined or null)
@@ -10,11 +10,11 @@ import { BlockId, Version } from "./responses";
  */
 export function assertSet<T>(value: T): T {
   if ((value as unknown) === undefined) {
-    throw new Error("Value must not be undefined");
+    throw new Error('Value must not be undefined');
   }
 
   if ((value as unknown) === null) {
-    throw new Error("Value must not be null");
+    throw new Error('Value must not be null');
   }
 
   return value;
@@ -28,8 +28,8 @@ export function assertSet<T>(value: T): T {
  */
 export function assertBoolean(value: boolean): boolean {
   assertSet(value);
-  if (typeof (value as unknown) !== "boolean") {
-    throw new Error("Value must be a boolean");
+  if (typeof (value as unknown) !== 'boolean') {
+    throw new Error('Value must be a boolean');
   }
   return value;
 }
@@ -42,8 +42,8 @@ export function assertBoolean(value: boolean): boolean {
  */
 export function assertString(value: string): string {
   assertSet(value);
-  if (typeof (value as unknown) !== "string") {
-    throw new Error("Value must be a string");
+  if (typeof (value as unknown) !== 'string') {
+    throw new Error('Value must be a string');
   }
   return value;
 }
@@ -56,8 +56,8 @@ export function assertString(value: string): string {
  */
 export function assertNumber(value: number): number {
   assertSet(value);
-  if (typeof (value as unknown) !== "number") {
-    throw new Error("Value must be a number");
+  if (typeof (value as unknown) !== 'number') {
+    throw new Error('Value must be a number');
   }
   return value;
 }
@@ -71,7 +71,7 @@ export function assertNumber(value: number): number {
 export function assertArray<T>(value: readonly T[]): readonly T[] {
   assertSet(value);
   if (!Array.isArray(value as unknown)) {
-    throw new Error("Value must be a an array");
+    throw new Error('Value must be a an array');
   }
   return value;
 }
@@ -85,15 +85,15 @@ export function assertArray<T>(value: readonly T[]): readonly T[] {
  */
 export function assertObject<T>(value: T): T {
   assertSet(value);
-  if (typeof (value as unknown) !== "object") {
-    throw new Error("Value must be an object");
+  if (typeof (value as unknown) !== 'object') {
+    throw new Error('Value must be an object');
   }
 
   // Exclude special kind of objects like Array, Date or Uint8Array
   // Object.prototype.toString() returns a specified value:
   // http://www.ecma-international.org/ecma-262/7.0/index.html#sec-object.prototype.tostring
-  if (Object.prototype.toString.call(value) !== "[object Object]") {
-    throw new Error("Value must be a simple object");
+  if (Object.prototype.toString.call(value) !== '[object Object]') {
+    throw new Error('Value must be a simple object');
   }
 
   return value;
@@ -114,25 +114,30 @@ interface Lengther {
 export function assertNotEmpty<T>(value: T): T {
   assertSet(value);
 
-  if (typeof value === "number" && value === 0) {
-    throw new Error("must provide a non-zero value");
+  if (typeof value === 'number' && value === 0) {
+    throw new Error('must provide a non-zero value');
   } else if ((value as any as Lengther).length === 0) {
-    throw new Error("must provide a non-empty value");
+    throw new Error('must provide a non-empty value');
   }
   return value;
 }
 
 // may will run the transform if value is defined, otherwise returns undefined
-export function may<T, U>(transform: (val: T) => U, value: T | null | undefined): U | undefined {
+export function may<T, U>(
+  transform: (val: T) => U,
+  value: T | null | undefined
+): U | undefined {
   return value === undefined || value === null ? undefined : transform(value);
 }
 
-export function dictionaryToStringMap(obj: Record<string, unknown>): Map<string, string> {
+export function dictionaryToStringMap(
+  obj: Record<string, unknown>
+): Map<string, string> {
   const out = new Map<string, string>();
   for (const key of Object.keys(obj)) {
     const value = obj[key];
-    if (typeof value !== "string") {
-      throw new Error("Found dictionary value of type other than string");
+    if (typeof value !== 'string') {
+      throw new Error('Found dictionary value of type other than string');
     }
     out.set(key, value);
   }
@@ -161,24 +166,33 @@ export function encodeUvarint(n: number): Uint8Array {
 export function encodeTime(time: ReadonlyDateWithNanoseconds): Uint8Array {
   const milliseconds = time.getTime();
   const seconds = Math.floor(milliseconds / 1000);
-  const secondsArray = seconds ? [0x08, ...encodeUvarint(seconds)] : new Uint8Array();
+  const secondsArray = seconds
+    ? [0x08, ...encodeUvarint(seconds)]
+    : new Uint8Array();
   const nanoseconds = (time.nanoseconds || 0) + (milliseconds % 1000) * 1e6;
-  const nanosecondsArray = nanoseconds ? [0x10, ...encodeUvarint(nanoseconds)] : new Uint8Array();
+  const nanosecondsArray = nanoseconds
+    ? [0x10, ...encodeUvarint(nanoseconds)]
+    : new Uint8Array();
   return Uint8Array.from([...secondsArray, ...nanosecondsArray]);
 }
 
 // See https://github.com/tendermint/go-amino/blob/v0.15.0/encoder.go#L180-L187
 export function encodeBytes(bytes: Uint8Array): Uint8Array {
   // Since we're only dealing with short byte arrays we don't need a full VarBuffer implementation yet
-  if (bytes.length >= 0x80) throw new Error("Not implemented for byte arrays of length 128 or more");
-  return bytes.length ? Uint8Array.from([bytes.length, ...bytes]) : new Uint8Array();
+  if (bytes.length >= 0x80)
+    throw new Error('Not implemented for byte arrays of length 128 or more');
+  return bytes.length
+    ? Uint8Array.from([bytes.length, ...bytes])
+    : new Uint8Array();
 }
 
 export function encodeVersion(version: Version): Uint8Array {
   const blockArray = version.block
     ? Uint8Array.from([0x08, ...encodeUvarint(version.block)])
     : new Uint8Array();
-  const appArray = version.app ? Uint8Array.from([0x10, ...encodeUvarint(version.app)]) : new Uint8Array();
+  const appArray = version.app
+    ? Uint8Array.from([0x10, ...encodeUvarint(version.app)])
+    : new Uint8Array();
   return Uint8Array.from([...blockArray, ...appArray]);
 }
