--- packages/core/src/lib/tendermintRpc/index.ts
@@ -1,106 +1,0 @@
-export {
-  pubkeyToAddress,
-  pubkeyToRawAddress,
-  rawEd25519PubkeyToRawAddress,
-  rawSecp256k1PubkeyToRawAddress,
-} from './addresses';
-export {
-  DateTime,
-  fromRfc3339WithNanoseconds,
-  fromSeconds,
-  ReadonlyDateWithNanoseconds,
-  toRfc3339WithNanoseconds,
-  toSeconds,
-} from './dates';
-// The public Tendermint34Client.create constructor allows manually choosing an RpcClient.
-// This is currently the only way to switch to the HttpBatchClient (which may become default at some point).
-// Due to this API, we make RPC client implementations public.
-export {
-  HttpBatchClient,
-  HttpBatchClientOptions,
-  HttpClient,
-  HttpEndpoint, // This type is part of the Tendermint34Client.connect API
-  RpcClient, // Interface type in Tendermint34Client.create
-  WebsocketClient,
-} from './rpcclients';
-export {
-  AbciInfoRequest,
-  AbciInfoResponse,
-  AbciQueryParams,
-  AbciQueryRequest,
-  AbciQueryResponse,
-  Attribute,
-  Block,
-  BlockchainRequest,
-  BlockchainResponse,
-  BlockGossipParams,
-  BlockId,
-  BlockMeta,
-  BlockParams,
-  BlockRequest,
-  BlockResponse,
-  BlockResultsRequest,
-  BlockResultsResponse,
-  BroadcastTxAsyncResponse,
-  BroadcastTxCommitResponse,
-  broadcastTxCommitSuccess,
-  BroadcastTxParams,
-  BroadcastTxRequest,
-  BroadcastTxSyncResponse,
-  broadcastTxSyncSuccess,
-  Commit,
-  CommitRequest,
-  CommitResponse,
-  ConsensusParams,
-  Event,
-  Evidence,
-  EvidenceParams,
-  GenesisRequest,
-  GenesisResponse,
-  Header,
-  HealthRequest,
-  HealthResponse,
-  Method,
-  NewBlockEvent,
-  NewBlockHeaderEvent,
-  NodeInfo,
-  NumUnconfirmedTxsRequest,
-  NumUnconfirmedTxsResponse,
-  ProofOp,
-  QueryProof,
-  QueryTag,
-  Request,
-  Response,
-  StatusRequest,
-  StatusResponse,
-  SubscriptionEventType,
-  SyncInfo,
-  TxData,
-  TxEvent,
-  TxParams,
-  TxProof,
-  TxRequest,
-  TxResponse,
-  TxSearchParams,
-  TxSearchRequest,
-  TxSearchResponse,
-  TxSizeParams,
-  Validator,
-  ValidatorsParams,
-  ValidatorsRequest,
-  ValidatorsResponse,
-  Version,
-  Vote,
-  VoteType,
-} from './tendermint34';
-export * as tendermint34 from './tendermint34';
-export { Tendermint34Client } from './tendermint34';
-export * as tendermint35 from './tendermint35';
-export { Tendermint35Client } from './tendermint35';
-export {
-  BlockIdFlag,
-  CommitSignature,
-  ValidatorEd25519Pubkey,
-  ValidatorPubkey,
-  ValidatorSecp256k1Pubkey,
-} from './types';
