--- packages/core/src/lib/tendermintRpc/rpcclients/index.ts
@@ -1,6 +1,0 @@
-// This folder contains Tendermint-specific RPC clients
-
-export { HttpBatchClient, HttpBatchClientOptions } from "./httpbatchclient";
-export { HttpClient, HttpEndpoint } from "./httpclient";
-export { instanceOfRpcStreamingClient, RpcClient, RpcStreamingClient, SubscriptionEvent } from "./rpcclient";
-export { WebsocketClient } from "./websocketclient";
