--- packages/core/src/lib/tendermintRpc/tendermint34/adaptor/index.ts
@@ -1,13 +1,0 @@
-import { hashBlock, hashTx } from "../hasher";
-import { Params } from "./requests";
-import { Responses } from "./responses";
-import { Adaptor } from "./types";
-
-export { Decoder, Encoder, Params, Responses } from "./types";
-
-export const adaptor34: Adaptor = {
-  params: Params,
-  responses: Responses,
-  hashTx: hashTx,
-  hashBlock: hashBlock,
-};
