--- packages/core/src/lib/tendermintRpc/tendermint34/index.ts
@@ -1,79 +1,0 @@
-// Note: all exports in this module are publicly available via
-// `import { tendermint34 } from "@cosmjs/tendermint-rpc"`
-
-export {
-  AbciInfoRequest,
-  AbciQueryParams,
-  AbciQueryRequest,
-  BlockchainRequest,
-  BlockRequest,
-  BlockResultsRequest,
-  BlockSearchParams,
-  BlockSearchRequest,
-  BroadcastTxParams,
-  BroadcastTxRequest,
-  CommitRequest,
-  GenesisRequest,
-  HealthRequest,
-  Method,
-  NumUnconfirmedTxsRequest,
-  QueryTag,
-  Request,
-  StatusRequest,
-  SubscriptionEventType,
-  TxParams,
-  TxRequest,
-  TxSearchParams,
-  TxSearchRequest,
-  ValidatorsParams,
-  ValidatorsRequest,
-} from "./requests";
-export {
-  AbciInfoResponse,
-  AbciQueryResponse,
-  Attribute,
-  Block,
-  BlockchainResponse,
-  BlockGossipParams,
-  BlockId,
-  BlockMeta,
-  BlockParams,
-  BlockResponse,
-  BlockResultsResponse,
-  BlockSearchResponse,
-  BroadcastTxAsyncResponse,
-  BroadcastTxCommitResponse,
-  broadcastTxCommitSuccess,
-  BroadcastTxSyncResponse,
-  broadcastTxSyncSuccess,
-  Commit,
-  CommitResponse,
-  ConsensusParams,
-  Event,
-  Evidence,
-  EvidenceParams,
-  GenesisResponse,
-  Header,
-  HealthResponse,
-  NewBlockEvent,
-  NewBlockHeaderEvent,
-  NodeInfo,
-  NumUnconfirmedTxsResponse,
-  ProofOp,
-  QueryProof,
-  Response,
-  StatusResponse,
-  SyncInfo,
-  TxData,
-  TxEvent,
-  TxProof,
-  TxResponse,
-  TxSearchResponse,
-  TxSizeParams,
-  Validator,
-  ValidatorsResponse,
-  Version,
-  Vote,
-  VoteType,
-} from "./responses";
-export { Tendermint34Client } from "./tendermint34client";
